# api/docs/swagger.yaml
basePath: /
definitions:
  api.AdminInstanceResponse:
    properties:
      challenge_name:
        type: string
      container_id:
        type: string
      created_at:
        type: string
      deployment_type:
        type: string
      expires_at:
        type: string
      hosted_address:
        type: string
      instance_id:
        type: string
      port:
        type: integer
      ttl_seconds:
        type: integer
      user_id:
        type: string
      username:
        type: string
    type: object
...
  api.HTTPErrorResp:
    properties:
      error:
        example: Error occured while veifying the challenge.
        type: string
    type: object
...
  api.TagInfoResp:
    properties:
      tags:
        items:
          type: string
        type: array
    type: object
...
  api.UserResp:
    properties:
      challenges:
        items:
          $ref: '#/definitions/api.ChallengeSolveResp'
        type: array
      email:
        example: fristonio@gmail.com
        type: string
      id:
        example: 5
        type: integer
      rank:
        example: 15
        type: integer
      role:
        example: author
        type: string
      score:
        example: 750
        type: integer
      status:
        example: 0
        type: integer
      username:
        example: CTF is live now!
        type: string
    type: object
...
info:
  contact:
    email: contact@sdslabs.co.in
    name: SDSLabs
    url: https://chat.sdslabs.co
  description: Authenticated API for Beast CTF challenge deployment and competition
    services.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Beast API
  version: "0.2"
paths:
  /api/admin/freezeLeaderboard:
    post:
      consumes:
      - application/json
      description: freezes the user leaderboard on demand.
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Freeze user leaderboard
      tags:
      - info
  /api/admin/instances:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.AdminInstanceResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all active instances
      tags:
      - admin
  /api/admin/instances/{instance_id}:
    delete:
      parameters:
      - description: Instance ID
        in: path
        name: instance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      security:
      - ApiKeyAuth: []
      summary: Delete an instance by ID
      tags:
      - admin
    get:
      parameters:
      - description: Instance ID
        in: path
        name: instance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AdminInstanceResponse'
      security:
      - ApiKeyAuth: []
      summary: Get an instance by ID
      tags:
      - admin
  /api/admin/instances/challenge/{challenge_name}:
    delete:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      security:
      - ApiKeyAuth: []
      summary: Delete all instances for a challenge
      tags:
      - admin
  /api/admin/instances/user/{user_id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      security:
      - ApiKeyAuth: []
      summary: Delete all instances owned by a user
      tags:
      - admin
  /api/admin/leaderboard:
    get:
      consumes:
      - application/json
      description: Returns admin leaderboard of all users
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Returns admin leaderboard
      tags:
      - info
  /api/admin/statistics:
    get:
      consumes:
      - application/json
      description: returns statistics of users in competition (currently limited to
        ban/unban status of users)
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UsersStatisticsResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: statistics of users in competition
      tags:
      - info
  /api/admin/submissions:
    get:
      consumes:
      - application/json
      description: Handles submissions made by the user
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SubmissionResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Handles submissions made by the user
      tags:
      - info
  /api/admin/unfreezeLeaderboard:
    post:
      consumes:
      - application/json
      description: unfreezes the user leaderboard on demand.
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Unfreeze user leaderboard
      tags:
      - info
  /api/admin/users/{action}/{id}:
    post:
      consumes:
      - application/json
      description: Ban/Unban/Hide/Unhide a user based on his user id. This operation
        can only be done by admins
      parameters:
      - description: Action to perform Ban/Unban/Hide/Unhide
        in: query
        name: action
        required: true
        type: string
      - description: Id of user
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChallengeStatusResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      summary: Ban/Unban/Hide/Unhide a user based on his id and the action provided.
      tags:
      - admin
  /api/config/challenge-info:
    post:
      consumes:
      - application/json
      description: Updates challenge info in the database, located at ~/.beast/beast.db.
      parameters:
      - description: Challenge Name
        in: formData
        name: name
        required: true
        type: string
      - description: Challenge's description
        in: formData
        name: desc
        type: string
      - description: Challenge's points
        in: formData
        name: points
        type: string
      - description: Challenge's flag
        in: formData
        name: flag
        type: string
      - description: Challenge's tags
        in: formData
        name: tags
        type: string
      - description: Challenge's ports
        in: formData
        name: ports
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Updates challenge info in the database, located at ~/.beast/beast.db.
      tags:
      - config
  /api/info/challenge/{name}:
    get:
      consumes:
      - application/json
      description: Returns all information about the challenges by the challenge name.
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Name of challenge
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Challenge'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Returns all information about the challenges.
      tags:
      - info
  /api/info/challenges:
    get:
      consumes:
      - application/json
      description: Returns information about all the challenges present in the database
        with and without filters.
      parameters:
      - description: Filter parameter by which challenges are filtered
        in: query
        name: filter
        type: string
      - description: Value of filtered parameter
        in: query
        name: value
        type: string
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ChallengeMetadata'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Returns metadata about all challenges with and without filters.
      tags:
      - info
  /api/info/competition-info:
    get:
      consumes:
      - application/json
      description: returns various information about the competition which are used
        to control competition
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CompetitionInfoResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: returns competition info
      tags:
      - info
  /api/info/download:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: The name of the challenge to get the logs for.
        in: query
        name: challenge
        type: string
      - description: The name of the static asset requested.
        in: query
        name: asset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LogsInfoResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      tags:
      - info
  /api/info/hint/{hintID}:
    get:
      parameters:
      - description: Hint ID
        in: path
        name: hintID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HintResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Read or purchase a challenge hint
      tags:
      - info
    post:
      parameters:
      - description: Hint ID
        in: path
        name: hintID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HintResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Read or purchase a challenge hint
      tags:
      - info
  /api/info/images/available:
    get:
      consumes:
      - application/json
      description: Returns all the available base images  which can be used for challenge
        creation as the base OS for challenge.
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AvailableImagesResp'
      summary: Gives all the base images that can be used while creating a beast challenge,
        this is a constant specified in beast global config
      tags:
      - info
  /api/info/leaderboard:
    get:
      consumes:
      - application/json
      description: Returns leaderboard of all users
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Returns leaderboard
      tags:
      - info
  /api/info/leaderboard-graph:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.UserLeaderboardResp'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Return leaderboard score history
      tags:
      - info
  /api/info/submissions/challenge/{challenge_id}:
    get:
      consumes:
      - application/json
      description: Returns all user attempts for a given challenge.
      parameters:
      - description: Challenge ID
        in: path
        name: challenge_id
        required: true
        type: integer
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.UserSolveResp'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Get challenge attempts
      tags:
      - info
  /api/info/submissions/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Returns all submissions for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.SubmissionResp'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Get submissions by user
      tags:
      - info
  /api/info/tags:
    get:
      consumes:
      - application/json
      description: returns all unique tags
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TagInfoResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: returns all tags
      tags:
      - info
  /api/info/user/{username}:
    get:
      consumes:
      - application/json
      description: Returns user info based on userId
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        type: integer
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Returns user info
      tags:
      - info
  /api/info/usercount:
    get:
      consumes:
      - application/json
      description: Returns the number of users in the database with role=contestant
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserCountResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Returns the number of users in the database with role=contestant
      tags:
      - info
  /api/info/users:
    get:
      consumes:
      - application/json
      description: Returns all available user's info
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sort by username or score
        in: query
        name: sort
        type: string
      - description: 'Score order: asc or desc'
        in: query
        name: order
        type: string
      - description: Filter by banned, active, or hidden
        in: query
        name: filter
        type: string
      - description: 'Response format: json or csv'
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.UsersResp'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Returns all user's info
      tags:
      - info
  /api/instances:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.InstanceResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List the current user's instances
      tags:
      - instances
  /api/instances/{challenge_name}:
    delete:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      security:
      - ApiKeyAuth: []
      summary: Delete the current user's challenge instance
      tags:
      - instances
    get:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InstanceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Get the current user's challenge instance
      tags:
      - instances
  /api/instances/{challenge_name}/extend:
    post:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      - description: Requested extension in seconds
        in: formData
        name: seconds
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InstanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Extend the current user's challenge instance
      tags:
      - instances
  /api/instances/{challenge_name}/spawn:
    post:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InstanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Spawn a per-user challenge instance
      tags:
      - instances
  /api/manage/challenge/:
    post:
      consumes:
      - application/json
      description: Handles challenge management routes with actions which includes
        - DEPLOY, UNDEPLOY, PURGE.
      parameters:
      - description: Name of the challenge to be managed, here name is the unique
          identifier for challenge
        in: query
        name: name
        required: true
        type: string
      - description: Action for the challenge
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      summary: Handles challenge management actions.
      tags:
      - manage
  /api/manage/challenge/{name}/exec:
    post:
      consumes:
      - application/json
      parameters:
      - description: Challenge name
        in: path
        name: name
        required: true
        type: string
      - description: Bounded exec request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ExecChallengeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ExecChallengeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Execute an argument-vector command in an owned challenge container
      tags:
      - manage
  /api/manage/challenge/multiple/:
    post:
      consumes:
      - application/json
      description: Handles challenge management routes with actions which includes
        - DEPLOY, UNDEPLOY, PURGE of multiple challenges.
      parameters:
      - description: Name of the challenge to be managed, here name is the unique
          identifier for challenges seperated by a comma
        in: query
        name: name
        required: true
        type: string
      - description: Action for the challenge
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      summary: Handles multiple challenge management actions.
  /api/manage/challenge/upload:
    post:
      consumes:
      - application/json
      description: Handles the challenge management from a challenge in zip file.
        Currently prepare the zip file
      parameters:
      - description: .zip file to be uploaded to fetch challenge info
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChallengePreviewResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Unzip and fetch info from beast.toml file in challenge
      tags:
      - manage
...
  /api/notification/stream:
    get:
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Stream server-sent notifications
      tags:
      - notification
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"