# api/router.go
package api

import (
	"io"
	"net/http"
	"path/filepath"
	"time"

	"github.com/gin-contrib/cors"
	"github.com/gin-contrib/static"
	"github.com/gin-gonic/gin"
	"github.com/sdslabs/beastv4/core"
	"github.com/sdslabs/beastv4/core/config"
	log "github.com/sirupsen/logrus"
)

func dummyHandler(c *gin.Context) {
	c.JSON(http.StatusOK, HTTPPlainResp{
		Message: WIP_TEXT,
	})
}

const maxAPIRequestBytes int64 = 2 << 20

func limitRequestBody(c *gin.Context) {
	limit := maxAPIRequestBytes
	if c.Request.URL.Path == "/api/manage/challenge/upload" {
		limit = maxChallengeUploadRequestBytes
	}
	if c.Request.ContentLength > limit {
		c.AbortWithStatusJSON(http.StatusRequestEntityTooLarge, HTTPErrorResp{Error: "Request body is too large"})
		return
	}
	c.Request.Body = http.MaxBytesReader(c.Writer, c.Request.Body, limit)
	c.Next()
}

func initGinRouter() *gin.Engine {
	router := gin.New()
	router.Use(gin.CustomRecoveryWithWriter(io.Discard, func(c *gin.Context, _ interface{}) {
		log.Error("request handler panic recovered")
		c.AbortWithStatusJSON(http.StatusInternalServerError, HTTPErrorResp{Error: "Internal server error"})
	}))

	if len(config.Cfg.ServerConfig.AllowedOrigins) > 0 {
		corsConfig := cors.Config{
			AllowOrigins:     config.Cfg.ServerConfig.AllowedOrigins,
			AllowMethods:     []string{"GET", "POST", "PUT", "PATCH", "DELETE"},
			AllowHeaders:     []string{"Origin", "Content-Length", "Content-Type", "Authorization"},
			AllowCredentials: false,
			MaxAge:           12 * time.Hour,
		}
		router.Use(cors.New(corsConfig))
	}
	router.Use(limitRequestBody)
	router.GET("/dummy", dummyHandler)
	// Authorization routes group
	authGroup := router.Group("/auth")
	{
		authGroup.POST("/register", register)
		authGroup.POST("/login", login)
		authGroup.POST("/reset-password", resetPasswordAuthorize, resetPasswordHandler)
		authGroup.POST("/send-otp", sendOTPHandler)
		authGroup.POST("/verify-otp", verifyOTPHandler)
		authGroup.POST("/send-otp-forget", sendOTPForForgetHandler)
		authGroup.POST("/verify-otp-forget", verifyOTPForForgetHandler)
	}

	// For serving static files
	router.Use(static.Serve("/api/info/logo", static.LocalFile(
		filepath.Join(core.BEAST_GLOBAL_DIR, core.BEAST_ASSETS_DIR, core.BEAST_LOGO_DIR),
		false)),
	)
	router.GET("/api/info/competition-info", competitionInfoHandler)
	router.GET("/api/info/download", serveAssets)

	// API routes group
	apiGroup := router.Group("/api", authorize)
	{
		// Deploy route group
		manageGroup := apiGroup.Group("/manage", managerAuthorize)
		{
			manageGroup.POST("/deploy/local/", adminAuthorize, deployLocalChallengeHandler)
			manageGroup.POST("/challenge/", manageChallengeHandler)
			manageGroup.POST("/challenge/multiple/", adminAuthorize, manageMultipleChallengeHandlerNameBased)
			manageGroup.POST("/multiple/:action", adminAuthorize, manageMultipleChallengeHandlerTagBased)
			manageGroup.POST("/static/:action", adminAuthorize, beastStaticContentHandler)
			manageGroup.POST("/commit/", commitChallenge)
			manageGroup.POST("/challenge/verify", verifyHandler)
			manageGroup.POST("/schedule/:action", adminAuthorize, manageScheduledAction)
			manageGroup.POST("/challenge/upload", manageUploadHandler)
			manageGroup.POST("/challenge/validateflag", validateFlagHandler)
			manageGroup.GET("/logs", challengeLogsHandler)
			manageGroup.POST("/challenge/:name/exec", execChallengeHandler)
		}

		// Status route group
		statusGroup := apiGroup.Group("/status")
		{
			statusGroup.GET("/challenge/:name", challengeStatusHandler)
			statusGroup.GET("/all", statusHandler)
			statusGroup.GET("/all/:filter", statusHandler)
		}

		// Info route group
		infoGroup := apiGroup.Group("/info")
		{
			infoGroup.GET("/challenge/:name", challengeInfoHandler)
			infoGroup.GET("/challenges", challengesMetadataHandler)
			infoGroup.GET("/images/available", availableImagesHandler)
			// infoGroup.GET("/ports/used", usedPortsInfoHandler)
			infoGroup.GET("/user/:username", userInfoHandler)
			infoGroup.GET("/users", getAllUsersInfoHandler)
			infoGroup.GET("/leaderboard", getLeaderboardHandler)
			infoGroup.GET("/leaderboard-graph", getLeaderboardGraphHandler)
			infoGroup.GET("/usercount", getUserCountHandler)
			infoGroup.GET("/submissions/challenge/:challenge_id", getChallengeAttempts)
			infoGroup.GET("/submissions/user/:user_id", getUserAttempts)
			infoGroup.GET("/tags", tagHandler)
			infoGroup.GET("/hint/:hintID", hintHandler)
			infoGroup.POST("/hint/:hintID", hintHandler)
		}

		// Notification route group
		notificationGroup := apiGroup.Group("/notification")
		{
			notificationGroup.GET("/available", availableNotificationHandler)
			notificationGroup.POST("/add", adminAuthorize, addNotification)
			notificationGroup.PUT("/update", adminAuthorize, updateNotifications)
			notificationGroup.DELETE("/delete", adminAuthorize, removeNotification)
			notificationGroup.GET("/stream", streamNotification)

		}

		remoteGroup := apiGroup.Group("/remote", adminAuthorize)
		{
			remoteGroup.POST("/sync", syncBeastGitRemote)
			remoteGroup.POST("/reset", resetBeastGitRemote)
		}

		configGroup := apiGroup.Group("/config", adminAuthorize)
		{
			configGroup.POST("/challenge-info", updateChallengeInfoHandler)
		}

		submitGroup := apiGroup.Group("/submit")
		{
			submitGroup.POST("/challenge", submitFlagHandler)
		}

		adminPanelGroup := apiGroup.Group("/admin", adminAuthorize)
		{
			adminPanelGroup.POST("/users/:action/:id", userActionHandler)
			adminPanelGroup.GET("/statistics", getUsersStatisticsHandler)
			adminPanelGroup.GET("/leaderboard", adminLeaderboardHandler)
			adminPanelGroup.POST("/freezeLeaderboard", freezeLeaderboardHandler)
			adminPanelGroup.POST("/unfreezeLeaderboard", unfreezeLeaderboardHandler)
			adminPanelGroup.GET("/submissions", submissionsHandler)

			adminPanelGroup.GET("/instances", adminGetAllInstancesHandler)
			adminPanelGroup.GET("/instances/:instance_id", adminGetInstanceHandler)
			adminPanelGroup.DELETE("/instances/:instance_id", adminKillInstanceHandler)
			adminPanelGroup.DELETE("/instances/user/:user_id", adminKillUserInstancesHandler)
			adminPanelGroup.DELETE("/instances/challenge/:challenge_name", adminKillChallengeInstancesHandler)
		}

		instanceGroup := apiGroup.Group("/instances")
		{
			instanceGroup.GET("", getUserInstancesHandler)
			instanceGroup.GET("/:challenge_name", getUserInstanceHandler)
			instanceGroup.POST("/:challenge_name/spawn", spawnInstanceHandler)
			instanceGroup.POST("/:challenge_name/extend", extendInstanceHandler)
			instanceGroup.DELETE("/:challenge_name", killUserInstanceHandler)
		}
	}

	router.NoRoute(func(c *gin.Context) {
		c.JSON(404, gin.H{"code": "PAGE_NOT_FOUND", "message": "Page not found"})
	})

	return router
}